+++ src/decoder/aac_decoder.ml
@@ -159,8 +159,8 @@ let file_type filename =
         channels;
       Some { Frame.audio = channels; video = 0; midi = 0 })
 
-let file_decoder ~metadata:_ ~kind filename =
-  Decoder.opaque_file_decoder ~filename ~kind create_decoder
+let file_decoder ~metadata:_ ~ctype filename =
+  Decoder.opaque_file_decoder ~filename ~ctype create_decoder
 
 let () =
   Decoder.decoders#register "AAC"