+++ src/decoder/gstreamer_decoder.ml
@@ -265,7 +265,7 @@ let get_type ~channels filename =
 let file_decoder ~metadata:_ ~ctype filename =
   let channels = ctype.Frame.audio in
   let content_type = get_type ~channels filename in
-  create_file_decoder filename content_type kind
+  create_file_decoder filename content_type ctype
 
 let () =
   Decoder.decoders#register "GSTREAMER"