+++ src/decoder/midi_decoder.ml
@@ -65,6 +65,7 @@ let () =
       file_extensions = (fun () -> Some ["mid"]);
       mime_types = (fun () -> None);
       file_type = (fun _ -> Some Frame.{ audio = 0; video = 0; midi = 1 });
-      file_decoder = Some (fun ~metadata:_ ~kind:_ filename -> decoder filename);
+      file_decoder =
+        Some (fun ~metadata:_ ~ctype:_ filename -> decoder filename);
       stream_decoder = None;
     }