+++ src/lang/builtins_request.ml
@@ -23,35 +23,6 @@
 open Lang_builtins
 
 let () =
-  add_builtin "request.create.raw" ~cat:Liq
-    ~descr:
-      "Create a raw request, for files that should not be decoded for \
-       streaming such as playlists. Creation may fail if there is no available \
-       RID, which cannot be detected currently: in that case one will obtain a \
-       request that will fail to be resolved."
-    [
-      ("indicators", Lang.list_t Lang.string_t, Some (Lang.list []), None);
-      ("persistent", Lang.bool_t, Some (Lang.bool false), None);
-      ("", Lang.string_t, None, None);
-    ]
-    (Lang.request_t
-       (Lang.frame_kind_t ~audio:Lang.zero_t ~video:Lang.zero_t
-          ~midi:Lang.zero_t))
-    (fun p ->
-      let indicators = List.assoc "indicators" p in
-      let persistent = Lang.to_bool (List.assoc "persistent" p) in
-      let initial = Lang.to_string (List.assoc "" p) in
-      let l = String.length initial in
-      let initial =
-        (* Remove trailing newline *)
-        if l > 0 && initial.[l - 1] = '\n' then String.sub initial 0 (l - 1)
-        else initial
-      in
-      let indicators = List.map Lang.to_string (Lang.to_list indicators) in
-      let indicators = List.map (fun x -> Request.indicator x) indicators in
-      Lang.request (Request.create_raw ~persistent ~indicators initial))
-
-let () =
   Lang.add_builtin "request.create" ~category:(string_of_category Liq)
     ~descr:
       "Create a request. Creation may fail if there is no available RID, which \
@@ -62,7 +33,7 @@ let () =
       ("persistent", Lang.bool_t, Some (Lang.bool false), None);
       ("", Lang.string_t, None, None);
     ]
-    (Lang.request_t (Lang.univ_t ()))
+    Lang.request_t
     (fun p ->
       let indicators = List.assoc "indicators" p in
       let persistent = Lang.to_bool (List.assoc "persistent" p) in
@@ -75,8 +46,7 @@ let () =
       in
       let indicators = List.map Lang.to_string (Lang.to_list indicators) in
       let indicators = List.map (fun x -> Request.indicator x) indicators in
-      (* TODO: we should ensure that the ctype is right by using Request.create instead... *)
-      Lang.request (Request.create_raw ~persistent ~indicators initial))
+      Lang.request (Request.create ~persistent ~indicators initial))
 
 let () =
   add_builtin "request.resolve" ~cat:Liq
@@ -85,32 +55,31 @@ let () =
         Lang.float_t,
         Some (Lang.float 30.),
         Some "Limit in seconds to the duration of the resolving." );
-      ("", Lang.request_t (Lang.univ_t ()), None, None);
+      ("", Lang.request_t, None, None);
     ]
     Lang.bool_t
     ~descr:
       "Resolve a request, i.e. attempt to get a valid local file. The \
        operation can take some time. Return true if the resolving was \
-       successful, false otherwise (timeout or invalid URI)."
+       successful, false otherwise (timeout or invalid URI). The request \
+       should not be decoded afterward: this is mostly useful to download \
+       files such as playlists, etc."
     (fun p ->
       let timeout = Lang.to_float (List.assoc "timeout" p) in
       let r = Lang.to_request (List.assoc "" p) in
       Lang.bool
-        (try Request.Resolved = Request.resolve r timeout with _ -> false))
+        ( try Request.Resolved = Request.resolve ~ctype:None r timeout
+          with _ -> false ))
 
 let () =
-  add_builtin "request.metadata" ~cat:Liq
-    [("", Lang.request_t (Lang.univ_t ()), None, None)]
-    Lang.metadata_t ~descr:"Get the metadata associated to a request."
-    (fun p ->
+  add_builtin "request.metadata" ~cat:Liq [("", Lang.request_t, None, None)]
+    Lang.metadata_t ~descr:"Get the metadata associated to a request." (fun p ->
       let r = Lang.to_request (List.assoc "" p) in
       Lang.metadata (Request.get_all_metadata r))
 
 let () =
-  add_builtin "request.log" ~cat:Liq
-    [("", Lang.request_t (Lang.univ_t ()), None, None)]
-    Lang.string_t ~descr:"Get log data associated to a request."
-    (fun p ->
+  add_builtin "request.log" ~cat:Liq [("", Lang.request_t, None, None)]
+    Lang.string_t ~descr:"Get log data associated to a request." (fun p ->
       let r = Lang.to_request (List.assoc "" p) in
       Lang.string (Request.string_of_log (Request.get_log r)))
 
@@ -119,18 +88,14 @@ let () =
     ~descr:
       "Check if a request is ready, i.e. is associated to a valid local file. \
        Unless the initial URI was such a file, a request has to be resolved \
-       before being ready."
-    [("", Lang.request_t (Lang.univ_t ()), None, None)]
-    Lang.bool_t
+       before being ready." [("", Lang.request_t, None, None)] Lang.bool_t
     (fun p ->
       let e = Lang.to_request (List.assoc "" p) in
       Lang.bool (Request.is_ready e))
 
 let () =
   add_builtin "request.uri" ~cat:Liq ~descr:"Initial URI of a request."
-    [("", Lang.request_t (Lang.univ_t ()), None, None)]
-    Lang.string_t
-    (fun p ->
+    [("", Lang.request_t, None, None)] Lang.string_t (fun p ->
       let r = Lang.to_request (List.assoc "" p) in
       Lang.string (Request.initial_uri r))
 
@@ -138,9 +103,7 @@ let () =
   add_builtin "request.filename" ~cat:Liq
     ~descr:
       "Return a valid local filename if the request is ready, and the empty \
-       string otherwise."
-    [("", Lang.request_t (Lang.univ_t ()), None, None)]
-    Lang.string_t
+       string otherwise." [("", Lang.request_t, None, None)] Lang.string_t
     (fun p ->
       let r = Lang.to_request (List.assoc "" p) in
       Lang.string (match Request.get_filename r with Some f -> f | None -> ""))
@@ -156,7 +119,7 @@ let () =
         Lang.bool_t,
         Some (Lang.bool false),
         Some "Destroy the request even if it is persistent." );
-      ("", Lang.request_t (Lang.univ_t ()), None, None);
+      ("", Lang.request_t, None, None);
     ]
     Lang.unit_t
     (fun p ->
@@ -177,8 +140,6 @@ let () =
 
 let () =
   add_builtin "request.id" ~cat:Liq ~descr:"Identifier of a request."
-    [("", Lang.request_t (Lang.univ_t ()), None, None)]
-    Lang.int_t
-    (fun p ->
+    [("", Lang.request_t, None, None)] Lang.int_t (fun p ->
       let r = Lang.to_request (List.assoc "" p) in
       Lang.int (Request.get_id r))