+++ src/operators/pipe.ml
@@ -44,7 +44,6 @@ class pipe ~kind ~replay_delay ~data_len ~process ~bufferize ~log_overfull ~max
   let log_ref = ref (fun _ -> ()) in
   let log x = !log_ref x in
   let log_error = ref (fun _ -> ()) in
-  let samplerate = Frame.audio_of_seconds 1. in
   let abg_max_len = Frame.audio_of_seconds max in
   let replay_delay = Frame.audio_of_seconds replay_delay in
   let resampler = Decoder_utils.samplerate_converter () in
@@ -64,75 +63,77 @@ class pipe ~kind ~replay_delay ~data_len ~process ~bufferize ~log_overfull ~max
 
     val mutable samplesize = 16
 
+    val mutable samplerate = Frame.audio_of_seconds 1.
+
     (* Filled in by wake_up. *)
     val mutable converter = fun _ -> assert false
 
     method private header =
       Bytes.unsafe_of_string
-        (Wav_aiff.wav_header ~channels:self#channels ~sample_rate ?len
-           ~sample_size:16 ())
+        (Wav_aiff.wav_header ~channels:self#channels ~sample_rate:samplerate
+           ?len ~sample_size:16 ())
 
     method private on_start push =
       Process_handler.really_write self#header push;
       `Continue
 
     method private on_stdout pull =
-          if not !header_read then (
-      let wav = Wav_aiff.read_header Wav_aiff.callback_ops pull in
-      header_read := true;
-      Tutils.mutexify mutex
-        (fun () ->
-          if Wav_aiff.channels wav <> channels then
-            failwith "Invalid channels from pipe process!";
-          samplesize := Wav_aiff.sample_size wav;
-          samplerate := Wav_aiff.sample_rate wav;
-          converter :=
-            Decoder_utils.from_iff ~format:`Wav ~channels
-              ~samplesize:!samplesize)
-        ();
-      `Reschedule Tutils.Non_blocking )
-    else (
-      let len = pull bytes 0 Utils.pagesize in
-      let data = !converter (Bytes.unsafe_to_string (Bytes.sub bytes 0 len)) in
-      let data = resampler ~samplerate:!samplerate data in
-      let len = Audio.length data in
-      let buffered = Generator.length abg in
-      Generator.put_audio abg data 0 len;
-      let to_replay =
+      if not !header_read then (
+        let wav = Wav_aiff.read_header Wav_aiff.callback_ops pull in
+        header_read := true;
         Tutils.mutexify mutex
           (fun () ->
-            let pending = !replay_pending in
-            let to_replay, pending =
-              List.fold_left
-                (fun ((pos, b), cur) (pos', b') ->
-                  if pos' + len > replay_delay then (
-                    if pos > 0 then
-                      log
-                        "Cannot replay multiple element at once.. Picking up \
-                         the most recent";
-                    if pos > 0 && pos < pos' then ((pos, b), cur)
-                    else ((pos', b'), cur) )
-                  else ((pos, b), (pos' + len, b') :: cur))
-                ((-1, `Nothing), [])
-                pending
-            in
-            replay_pending := pending;
-            to_replay)
-          ()
-      in
-      begin
-        match to_replay with
-        | -1, _ -> ()
-        | _, `Break_and_metadata m ->
-            Generator.add_metadata abg m;
-            Generator.add_break abg
-        | _, `Metadata m -> Generator.add_metadata abg m
-        | _, `Break -> Generator.add_break abg
-        | _ -> ()
-      end;
-      if abg_max_len < buffered + len then
-        `Delay (Frame.seconds_of_audio (buffered + len - abg_max_len))
-      else `Continue )
+            if Wav_aiff.channels wav <> self#channels then
+              failwith "Invalid channels from pipe process!";
+            samplesize <- Wav_aiff.sample_size wav;
+            samplerate <- Wav_aiff.sample_rate wav;
+            converter <-
+              Decoder_utils.from_iff ~format:`Wav ~channels:self#channels
+                ~samplesize)
+          ();
+        `Reschedule Tutils.Non_blocking )
+      else (
+        let len = pull bytes 0 Utils.pagesize in
+        let data = converter (Bytes.unsafe_to_string (Bytes.sub bytes 0 len)) in
+        let data = resampler ~samplerate data in
+        let len = Audio.length data in
+        let buffered = Generator.length abg in
+        Generator.put_audio abg data 0 len;
+        let to_replay =
+          Tutils.mutexify mutex
+            (fun () ->
+              let pending = !replay_pending in
+              let to_replay, pending =
+                List.fold_left
+                  (fun ((pos, b), cur) (pos', b') ->
+                    if pos' + len > replay_delay then (
+                      if pos > 0 then
+                        log
+                          "Cannot replay multiple element at once.. Picking up \
+                           the most recent";
+                      if pos > 0 && pos < pos' then ((pos, b), cur)
+                      else ((pos', b'), cur) )
+                    else ((pos, b), (pos' + len, b') :: cur))
+                  ((-1, `Nothing), [])
+                  pending
+              in
+              replay_pending := pending;
+              to_replay)
+            ()
+        in
+        begin
+          match to_replay with
+          | -1, _ -> ()
+          | _, `Break_and_metadata m ->
+              Generator.add_metadata abg m;
+              Generator.add_break abg
+          | _, `Metadata m -> Generator.add_metadata abg m
+          | _, `Break -> Generator.add_break abg
+          | _ -> ()
+        end;
+        if abg_max_len < buffered + len then
+          `Delay (Frame.seconds_of_audio (buffered + len - abg_max_len))
+        else `Continue )
 
     val mutable handler = None
 
@@ -210,29 +211,29 @@ class pipe ~kind ~replay_delay ~data_len ~process ~bufferize ~log_overfull ~max
 
     method private on_stop =
       Tutils.mutexify mutex (fun e ->
-        let ret = !next_stop in
-        next_stop := `Nothing;
-        header_read := false;
-        let should_restart =
-          match e with
-            | `Status s when s <> Unix.WEXITED 0 -> restart_on_error
-            | `Exception _ -> restart_on_error
-            | _ -> true
-        in
-        match (should_restart, ret) with
-          | false, _ -> false
-          | _, `Sleep -> false
-          | _, `Break_and_metadata m ->
-              Generator.add_metadata abg m;
-              Generator.add_break abg;
-              true
-          | _, `Metadata m ->
-              Generator.add_metadata abg m;
-              true
-          | _, `Break ->
-              Generator.add_break abg;
-              true
-          | _, `Nothing -> restart)
+          let ret = !next_stop in
+          next_stop := `Nothing;
+          header_read := false;
+          let should_restart =
+            match e with
+              | `Status s when s <> Unix.WEXITED 0 -> restart_on_error
+              | `Exception _ -> restart_on_error
+              | _ -> true
+          in
+          match (should_restart, ret) with
+            | false, _ -> false
+            | _, `Sleep -> false
+            | _, `Break_and_metadata m ->
+                Generator.add_metadata abg m;
+                Generator.add_break abg;
+                true
+            | _, `Metadata m ->
+                Generator.add_metadata abg m;
+                true
+            | _, `Break ->
+                Generator.add_break abg;
+                true
+            | _, `Nothing -> restart)
 
     method private slave_tick =
       (Clock.get source#clock)#end_tick;
@@ -249,8 +250,7 @@ class pipe ~kind ~replay_delay ~data_len ~process ~bufferize ~log_overfull ~max
     method wake_up a =
       super#wake_up a;
       converter <-
-        Rutils.create_from_iff ~format:`Wav ~channels:self#channels ~samplesize
-          ~audio_src_rate;
+        Decoder_utils.from_iff ~format:`Wav ~channels:self#channels ~samplesize;
 
       source#get_ready [(self :> source)];
       (* Now we can create the log function *)
@@ -258,9 +258,9 @@ class pipe ~kind ~replay_delay ~data_len ~process ~bufferize ~log_overfull ~max
       log_error := self#log#debug "%s";
       handler <-
         Some
-          (Process_handler.run ~on_stop ~on_start:self#on_start
+          (Process_handler.run ~on_stop:self#on_stop ~on_start:self#on_start
              ~on_stdout:self#on_stdout ~on_stdin:self#on_stdin
-             ~priority:Tutils.Blocking ~on_stderr ~log process)
+             ~priority:Tutils.Blocking ~on_stderr:self#on_stderr ~log process)
 
     method abort_track = source#abort_track