+++ CHANGES.md
@@ -86,6 +86,8 @@ Changed:
 - Allow crossfading for video (#1132, #1135).
 - Use getters for parameters of synthesizer sources (#1036).
 - Renamed `empty` to `fail`.
+- Requests are not typed anymore: their type is fixed at resolution time.
+- Deprecated `request.create.raw`, you should use `request.create` instead.
 
 Fixed: